% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freeze-universe.R
\name{snapshot_package_repository}
\alias{snapshot_package_repository}
\title{Snapshot a Package Repository}
\usage{
snapshot_package_repository(
  url,
  destdir = basename(url),
  type = c("source", "mac.binary", "win.binary"),
  r_version = NULL
)
}
\arguments{
\item{url}{URL to the CRAN-like repository.}

\item{destdir}{Folder where to save the archives.}

\item{type}{Type of package archives.}

\item{r_version}{R version}
}
\description{
Snapshot a Package Repository
}
\examples{
\dontrun{
snapshot_package_repository("https://jeroen.r-universe.dev", type = "source")
}
}
